Update installation instructions to install from GitHub repository
Due to the dependency on the `chirp` project, which is not available on PyPi, we are unable to publish version 2.0.3 of `wwara_chirp` package to PyPi. The `chirp` project is not available on PyPi so the dependency mist be installed direct from the kk7ds/chirp repository. This requires that v2.0.3 be installed from source at the tsayles/wwara_chirp repo.