WIP 2025-09-03 activaly editing update_constants.py
Search for "# TODO START_HERE"

Needs code to find the mock constants in
the chirp_common.py. refactor to open chirp_common.py once
and use ast to find the constants in the file. Need to impliment
a way to identify the changed constants, and then update the
target file with the new values.

Edits so far:
Refactor fetch_file method and enhance constant extraction in update_constants.py